backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
psycopg[binary]==3.2.6
pgvector==0.3.6
pydantic-settings==2.8.1
python-multipart==0.0.20
httpx==0.28.1
pypdf==5.4.0
numpy==2.2.4
scikit-learn==1.6.1
openai==1.70.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
email-validator==2.2.0
